Add TAP output for tests, use it to integrate with Meson
Instead of building a single run-tests executable, we build lots of
smaller tests which output TAP.  This is simple enough that I think
we can drop the µnit dependency, and it has a few advantages such as
built-in support for parallel tests and better integration with
tooling.

The µnit version will remain in place for now, but only work with
CMake.  Meson, OTOH, will only work with this.  I plan to remove
CMake support in the relatively near future, just need to get a few
remaining builds swiched over.